⚡️ :: readTextFile, Realm 구조 변경
